Revert "refactor(compiler): expose utility for creating CSS selectors from AST nodes (#52726)" (#53012)
This reverts commit adbea7befbc372c93b17c7d978356d1e2dce99bc.

PR Close #53012
